+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+v1.0.4 (2019-09-08)
+
+* Added GV11 for Sentry Mode
+* Added GV12 for drivers side climate control temperature (celsius only)
+* Added GV13 for passengers side climate control temperature (celsius only)
+* Added GV14 for outside temperature (celsius only)
+* Added cmd WINDOWS_VENT to vent all the windows (requires Tesla V10 firmware)
+* Added cmd WINDOWS_CLOSE to close all the windows (requires Tesla V10 firmware)
+* Added cmd TRUNK_OPEN to open the rear trunk
+* Added cmd FRUNK_OPEN to open the front trunk (frunk)
+* Added cmd HEATED_SEAT_LEVEL_DRIVER to set the level on the heated seat for the driver
+* Added cmd HEATED_SEAT_LEVEL_PASSENGER to set the level on the heated seat for the passenger
+* Added cmd HEATED_SEAT_LEVEL_REAR_LEFT to set the level on the heated seat for the rear left seat
+* Added cmd HEATED_SEAT_LEVEL_REAR_CENTER to set the level on the heated seat for the rear center seat
+* Added cmd HEATED_SEAT_LEVEL_REAR_RIGHT to set the level on the heated seat for the rear right seat
+* Added cmd SENTRY_MODE_ON to turn on Sentry Mode
+* Added cmd SENTRY_MODE_OFF to turn off Sentry Mode
+* Added cmd START_SOFTWARE_UPDATE to start the Tesla software update if one is available.
+* Added cmd MAX_DEFROST_ON to turn on the Max Defrost
+* Added cmd MAX_DEFROST_OFF to turn off the Max Defrost
+* Added cmd CLIMATE_TEMP_SETTING_DRIVER to set the climate control temperature for the drivers side (celsius only)
+* Added cmd CLIMATE_TEMP_SETTING_PASSENGER to set the climate control temperature for the passengers side (celsius only)
+* Added check for distance units. Will display KMs if the vehicle GUI is in KM/HR
+
 v1.0.3 (2019-06-14)
 * Added CLITEMP internal temperature
 * Added cmd CLIMATE_ON to start conditioning the vehicle